refactor: Move backend routes and models to separate directories  and change file structure in frontend
The backend routes and models were reorganized to improve code organization and maintainability. The `Items.js` route file was moved to the `items/router` directory, and the `Item.js` model file was moved to the `items/models` directory. This separation of concerns makes it easier to locate and manage related files.